+++ src/EventListener/Workflow/AddFormattedEntityNotificationTokensListener.php
@@ -15,10 +15,14 @@ declare(strict_types=1);
 
 namespace Netzmacht\ContaoWorkflowBundle\EventListener\Workflow;
 
+use Contao\DC_Table;
+use Contao\Model;
 use Netzmacht\Contao\Toolkit\Assertion\AssertionFailed;
 use Netzmacht\Contao\Toolkit\Dca\DcaManager;
 use Netzmacht\ContaoWorkflowBundle\PropertyAccess\PropertyAccessManager;
 use Netzmacht\ContaoWorkflowBundle\Workflow\Flow\Action\Notification\BuildNotificationTokensEvent;
+use Throwable;
+use function iterator_to_array;
 
 /**
  * Class AddFormattedEntityNotificationTokensListener enriches the notifications with formatted values.
@@ -73,10 +77,18 @@ final class AddFormattedEntityNotificationTokensListener
             return;
         }
 
-        foreach ($this->propertyAccessManager->provideAccess($entity) as $key => $value) {
+        $context = null;
+        if ($entity instanceof Model) {
+            $accessor = $this->propertyAccessManager->provideAccess($entity);
+            $context  = new DC_Table($event->getItem()->getEntityId()->getProviderName());
+
+            $context->activeRecord = (object) iterator_to_array($accessor->getIterator());
+        }
+
+        foreach ($accessor as $key => $value) {
             try {
-                $event->addToken('formatted_' . $key, $formatter->formatValue($key, $value));
-            } catch (\Exception $exception) {
+                $event->addToken('formatted_' . $key, $formatter->formatValue($key, $value, $context));
+            } catch (Throwable $exception) {
                 // Skip if any error occurs
             }
         }